+++ almoxarifado-v0.2.2/aplicacao/clinicas/src/main/java/ufc/npi/clinicas/controller/SaidaController.java
@@ -237,22 +237,6 @@ public class SaidaController {
 
 	}
 
-	@GetMapping(value = "saidaMaterial/remover/{idItemSaidaMaterial}")
-	public Response excluiAlocacaoItemMaterial(@PathVariable("idItemSaidaMaterial") Long idItemSaidaMaterial) {
-		ItemSaida itemSaida = itemSaidaService.buscarPorId(idItemSaidaMaterial);
-
-		if (itemSaida != null) {
-			if (itemSaida.getSaidaMaterial().getStatus().equals(Status.EM_ANDAMENTO)) {
-				itemSaidaService.excluirItemSaidaMaterial(idItemSaidaMaterial);
-				return new Response().withSuccessMessage(Constants.SAIDA_INCLUIR_MATERIAIS_SUCESSO_REMOVER);
-			}
-		}
-		else{
-			return new Response().withFailStatus().withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_REMOVER_NULL);
-		}
-		return new Response().withFailStatus().withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_REMOVER_ERRO);
-	}
-
 	@PostMapping("/api/itemSaida/adicionar")
 	public Response adicionarItemSaida(Integer quantidade, Integer idSaida, Integer idMaterial, Integer idEstoqueLote, Integer idSetor){
